% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexFitting.R
\name{flex_fit}
\alias{flex_fit}
\title{Easy curve fitting in R}
\usage{
flex_fit(y, x, logistic = FALSE, aug = FALSE)
}
\arguments{
\item{y}{A numerical vector}

\item{x}{A numerical vector}

\item{logistic}{a logical value indicating whether logistic formula should be
fitted.}

\item{aug}{a logical value indicating whether augment should be export}
}
\value{
flex_fit(): Returns a tibble or list
}
\description{

}
\examples{
# Compute a formula
flex_fit(mtcars$mpg,mtcars$disp)
}
