% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexFitting.R
\name{show_formulalist}
\alias{show_formulalist}
\title{Show the formula list}
\usage{
show_formulalist()
}
\value{
show_formulalist(): Returns a data.frame
}
\description{
Show the formula list
}
\examples{
show_formulalist()
}
